feat: refactored environments so that kuhn and all relevant data classes and types will be contained within itself. Started on main env
