+++ src/kuhn/datatypes.py
@@ -136,7 +136,7 @@ class HistoricalKuhn:
         self.rule_params = K.rule_params
         self.rule_params['betsizes'] = T([0.5,1.])
         self.state_params = K.state_params
-        self.state_params['stacksize'] = 3
+        self.state_params['stacksize'] = 5
 
 class BaseHoldem(object):
     def __init__(self):
